DSND-543: IT Payload files update
