+++ src/itest/resources/payload/input/alterations_to_order_input.json
@@ -5,7 +5,7 @@
                 "name": "H I D B"
             }
         ],
-        "id": "REU4N0IyMTI5MUQ0NDFGODVDQzdDNzQ3MjFBMzg5N0EyNjIzQ0YxMw==",
+        "id": "AbRiNTU3NjNjZWI1Y2YxMzkzYWY3MzQ0YzVlOTg4ZGVhZTBkYWI4Ng==",
         "charge_number": 2,
         "particulars": {
             "type": "short-particulars",
@@ -44,33 +44,33 @@
         ],
         "insolvency_cases": null,
         "links": {
-            "self": "/company/SC109614/charges/-1yobYByNCXBCrEPzoYi1CEAF_0="
-        }
+            "self": "/company/88124207/charges/AbRiNTU3NjNjZWI1Y2YxMzkzYWY3MzQ0YzVlOTg4ZGVhZTBkYWI4Ng=="
+        },
+        "status": "fully-satisfied",
+        "assets_ceased_released": null,
+        "acquired_on": null,
+        "delivered_on": "1989-03-15",
+        "resolved_on": null,
+        "covering_instrument_date": null,
+        "created_on": "1989-03-01",
+        "satisfied_on": "1998-03-24",
+        "classification": {
+            "type": "charge-description",
+            "description": "FLOATING CHARGE"
+        },
+        "secured_details": {
+            "type": "obligations-secured",
+            "description": "FLOATING CHARGE"
+        },
+        "scottish_alterations": {
+            "type": null,
+            "description": null,
+            "has_alterations_to_order": false,
+            "has_alterations_to_prohibitions": false,
+            "has_restricting_provisions": true
+        },
+        "more_than_four_persons_entitled": false
     },
-    "status": "fully-satisfied",
-    "assets_ceased_released": null,
-    "acquired_on": null,
-    "delivered_on": "1989-03-15",
-    "resolved_on": null,
-    "covering_instrument_date": null,
-    "created_on": "1989-03-01",
-    "satisfied_on": "1998-03-24",
-    "classification": {
-        "type": "charge-description",
-        "description": "FLOATING CHARGE"
-    },
-    "secured_details": {
-        "type": "obligations-secured",
-        "description": "FLOATING CHARGE"
-    },
-    "scottish_alterations": {
-        "type": null,
-        "description": null,
-        "has_alterations_to_order": false,
-        "has_alterations_to_prohibitions": false,
-        "has_restricting_provisions": true
-    },
-    "more_than_four_persons_entitled": false,
     "internal_data": {
         "delta_at": "2015-06-25T22:00:00.000Z",
         "updated_by": "test-partition_1-offset_1"
