+++ src/itest/resources/payload/output/Scott_alterations_output.json
@@ -1,20 +1,23 @@
 {
-    "_id" : "TEU4N0IyMTI5MUQ0NDFGODVDQzdDNzQ3MjFBMzg5N0EyNjIzQ0YxMw==",
-    "company_number" : "SC109614",
+    "_id" : "AbRiNTU3NjNjZWI1Y2YxMzkzYWY3MzQ0YzVlOTg4ZGVhZTBkYWI4Ng==",
+    "company_number" : "88124207",
     "data" : {
-        "etag" : "f9a3110b59ff5651966dc10dcecb148cbac8ac7c",
-        "id" : "TEU4N0IyMTI5MUQ0NDFGODVDQzdDNzQ3MjFBMzg5N0EyNjIzQ0YxMw==",
+        "etag" : "07473d46da1dc65cc4829d64c5f5c4465880ab55",
+        "id" : "AbRiNTU3NjNjZWI1Y2YxMzkzYWY3MzQ0YzVlOTg4ZGVhZTBkYWI4Ng==",
         "charge_code" : null,
-        "classification" : null,
+        "classification" : {
+            "type" : "charge-description",
+            "description" : "FLOATING CHARGE"
+        },
         "charge_number" : 2,
-        "status" : null,
+        "status" : "fully-satisfied",
         "assets_ceased_released" : null,
         "acquired_on" : null,
-        "delivered_on" : null,
+        "delivered_on" : "1989-03-15T00:00:00.000Z",
         "resolved_on" : null,
         "covering_instrument_date" : null,
-        "created_on" : null,
-        "satisfied_on" : null,
+        "created_on" : "1989-03-01T00:00:00.000Z",
+        "satisfied_on" : "1998-03-24T00:00:00.000Z",
         "particulars" : {
             "type" : "short-particulars",
             "description" : "UNDERTAKING AND ALL PROPERTY AND ASSETS PRESENT AND FUTURE OF THE COMPANY INCLUDING UNCALLED CAPITAL",
@@ -24,15 +27,24 @@
             "contains_negative_pledge" : false,
             "chargor_acting_as_bare_trustee" : false
         },
-        "secured_details" : null,
-        "scottish_alterations" : null,
-        "more_than_four_persons_entitled" : null,
-        "persons_entitled" : [ 
+        "secured_details" : {
+            "type" : "obligations-secured",
+            "description" : "FLOATING CHARGE"
+        },
+        "scottish_alterations" : {
+            "type" : null,
+            "description" : null,
+            "has_alterations_to_order" : true,
+            "has_alterations_to_prohibitions" : true,
+            "has_restricting_provisions" : true
+        },
+        "more_than_four_persons_entitled" : false,
+        "persons_entitled" : [
             {
                 "name" : "H I D B"
             }
         ],
-        "transactions" : [ 
+        "transactions" : [
             {
                 "filing_type" : "create-charge",
                 "transaction_id" : null,
@@ -46,7 +58,7 @@
         ],
         "insolvency_cases" : null,
         "links" : {
-            "self" : "/company/SC109614/charges/TEU4N0IyMTI5MUQ0NDFGODVDQzdDNzQ3MjFBMzg5N0EyNjIzQ0YxMw=="
+            "self" : "/company/88124207/charges/AbRiNTU3NjNjZWI1Y2YxMzkzYWY3MzQ0YzVlOTg4ZGVhZTBkYWI4Ng=="
         }
     },
     "updated" : {
