--- src/main/java/uk/gov/companieshouse/charges/data/config/MongoConfig.java
@@ -1,39 +1,0 @@
-package uk.gov.companieshouse.charges.data.config;
-
-import com.mongodb.ConnectionString;
-import com.mongodb.MongoClientSettings;
-import com.mongodb.client.MongoClient;
-import com.mongodb.client.MongoClients;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.mongodb.config.AbstractMongoClientConfiguration;
-
-@Configuration
-public class MongoConfig extends AbstractMongoClientConfiguration {
-
-    @Value("${spring.data.mongodb.name}")
-    private String databaseName;
-
-    @Value("${spring.data.mongodb.uri}")
-    private String databaseUri;
-
-
-    @Override
-    protected String getDatabaseName() {
-        return this.databaseName;
-    }
-
-    protected String getDatabaseUri() {
-        return this.databaseUri;
-    }
-
-    @Override
-    public MongoClient mongoClient() {
-        final ConnectionString connectionString =
-                new ConnectionString(getDatabaseUri());
-        final MongoClientSettings mongoClientSettings = MongoClientSettings.builder()
-                .applyConnectionString(connectionString)
-                .build();
-        return MongoClients.create(mongoClientSettings);
-    }
-}
