+++ src/main/java/uk/gov/companieshouse/charges/data/config/MongoTransactionConfig.java
@@ -1,18 +1,50 @@
 package uk.gov.companieshouse.charges.data.config;
 
+import com.mongodb.ConnectionString;
+import com.mongodb.MongoClientSettings;
+import com.mongodb.client.MongoClient;
+import com.mongodb.client.MongoClients;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.mongodb.MongoDatabaseFactory;
 import org.springframework.data.mongodb.MongoTransactionManager;
+import org.springframework.data.mongodb.config.AbstractMongoClientConfiguration;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 
 @ConditionalOnProperty(name = "mongodb.charges.transactions", havingValue = "true")
 @Configuration
 @EnableTransactionManagement
-public class MongoTransactionConfig {
+public class MongoTransactionConfig extends AbstractMongoClientConfiguration {
+
+    @Value("${spring.data.mongodb.name}")
+    private String databaseName;
+
+    @Value("${spring.data.mongodb.uri}")
+    private String databaseUri;
+
     @Bean
     MongoTransactionManager transactionManager(MongoDatabaseFactory dbFactory) {
         return new MongoTransactionManager(dbFactory);
+    }
+
+    @Override
+    protected String getDatabaseName() {
+        return this.databaseName;
+    }
+
+    protected String getDatabaseUri() {
+        return this.databaseUri;
+    }
+
+    @Override
+    public MongoClient mongoClient() {
+        final ConnectionString connectionString =
+                new ConnectionString(getDatabaseUri());
+        final MongoClientSettings mongoClientSettings = MongoClientSettings.builder()
+                .applyConnectionString(connectionString)
+                .build();
+        return MongoClients.create(mongoClientSettings);
     }
 }
